vendas 1

+++ backend/src/main/java/com/loja/model/Venda.java
@@ -1,15 +1,16 @@
 package com.loja.model;
 import java.util.List;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 public class Venda {
   private Cliente cliente;
   private List<Produto> produtos;
-  private Double valorTotal;
+  private BigDecimal valorTotal;
   private LocalDateTime dataHora;
   
   public Venda() {}
 
-  public Venda(Cliente cliente, List<Produto> produtos, Double valorTotal, LocalDateTime dataHora) {
+  public Venda(Cliente cliente, List<Produto> produtos, BigDecimal valorTotal, LocalDateTime dataHora) {
 	this.cliente = cliente;
 	this.produtos = produtos;
 	this.valorTotal = valorTotal;
@@ -32,11 +33,11 @@ public class Venda {
 	this.produtos = produtos;
   }
 
-  public Double getValorTotal() {
+  public BigDecimal getValorTotal() {
 	return valorTotal;
   }
 
-  public void setValorTotal(Double valorTotal) {
+  public void setValorTotal(BigDecimal valorTotal) {
 	this.valorTotal = valorTotal;
   }
